requirements.txt
numpy==1.23.4
pandas==1.5.1
Pillow==9.3.0
requests==2.28.1
scikit_learn==1.2.0
tensorflow==2.10.0
transformers==4.24.0
